+++ Client/my-story-illustrator-client/src/header/user_settings_button.js
