+++ Client/my-story-illustrator-client/src/index-backup.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import App from './App';
+
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <React>
+    <App />
+  </React>
+);
+
