+++ Client/my-story-illustrator-client/src/index.js
@@ -1,12 +1,25 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+
+import MyStories from './my_stories/my_stories';
+import Header from './header/header';
+
+import './global.css';
+import './my_stories/my_stories.css';
+
+const stories = [{Title : "testing title", CoverImageUri : "https://picsum.photos/200/300"}]
+
+const justSignedIn = false;
+const editingStory = false;
+
+const element = (
+    <body>
+        <Header justSignedIn = {justSignedIn} editingStory = {editingStory} />
+        <MyStories stories={stories} />
+    </body>
+
+)
 
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-  <React>
-    <App />
-  </React>
-);
-
+root.render(element);
