+++ Client/my-story-illustrator-client/src/my_stories/make_buttons_boring.css
