+++ Client/my-story-illustrator-client/src/my_stories/my_stories.css
