+++ Client/my-story-illustrator-client/src/my_stories/new_story_button.js
