+++ Client/my-story-illustrator-client/src/my_stories/story_cover_button.js
