+++ Client/my-story-illustrator-client/src/read_book/image_page.js
