+++ Client/my-story-illustrator-client/src/read_book/page_footer.js
